Request user all permissions interactively
Before this change not all permissions were requested, so the permissions has to be activated first. Also, some permissions were requested after they are needed, thus the application had to be reinitialized after grant such permissions.
After this change, all permissions are requested to the user before they are needed. So no functionality is started before the grant of the corresponding permission.