Background functionality improved.
The app request for not to be included in the battery optimizations, what can stop the app in the background otherwise. An info activity was added, in order to communicate the user about this issue. There are also instructions for Samsung devices, which uses a different behaviour and the user has to do some changes in the settings of the device manually.
Also, the activity can be instanced once. Before this change, each time the notification icon was pressed, a new activity was created, having the incorrect behaviour of when the user press the back button, a lot of main activities are dismissed. What is wrong.