Fixed multiple calls of the GPS state changes
Before this change, the GPS state callback object was called several times because the object wasn't unregistered whenever the application is finished. Since Android keeps the state a while in order to speed up the apps initialisation, the former registration remains while a new registration is performed. This behaviour mess the GNSS state handling up.

After this change, the GNSS state callback object is unregistered when the app is finishing.