Show info view on app startup
Before this change, the info view was shown using the menu only.

After this change, the info view is shown on app's startup. The info view includes also a checkbox, so the user can disable showing this on the startup. Also some instructions about the battery save mode of the device.