Deprecated method TextToSpeech.speak replaced
Before this change, the base code was using a deprecated overload of the TextToSpeech.speak method. This situation could lead to incompatibilities with future Android versions.

After this change, a current overload of the method is using. This method requires an utterance-id, so that one is given for each use of the method.